Track key platform files
This is in order to version key items that cause problems when rebuilding from scratch.
For Alexa, this notably includes the Alexa Skill id.
For Google Action, this notably includes the Account Linking settings.
TODO: We should be able to figure out a way to compliment the Jovo build with steps that inject the key values here, rather than rely on versioning some things and not others.